fix for optimize collect
