+++ native-engine/datafusion-ext-commons/src/io/mod.rs
@@ -19,10 +19,12 @@ use arrow::{
     datatypes::{DataType, SchemaRef},
     record_batch::RecordBatch,
 };
-pub use batch_serde::{read_array, read_data_type, read_scalar, write_array, write_data_type, write_scalar};
+pub use batch_serde::{read_array, read_data_type, write_array, write_data_type};
+pub use scalar_serde::{read_scalar, write_scalar};
 use datafusion::common::{cast::as_struct_array, Result};
 
 mod batch_serde;
+mod scalar_serde;
 
 pub fn write_one_batch<W: Write + Seek>(
     batch: &RecordBatch,
